Updated listings routes and functions. Completed comments and tradeRequest functions, and getListing handlebars.
Completed all of the functions for comments and traderRquest, added functions for error checking in helpers. Started the initial skeleton for the listings handlebars. Fixed the updateListing function and fixed some issues with listing Routes. **Have not error tested all functions from comments and tradeRequests, so some may not work properly!!